refactor: rename src folder to lib
