+++ lib/format-props.js
