+++ lib/generate-tag.js
