+++ lib/h.js
