+++ lib/sanitize.js
