+++ lib/stringify.js
