+++ lib/view.decorator.js
