+++ sonar-project.properties
@@ -4,8 +4,8 @@ sonar.projectKey=lite-jsx_core
 sonar.sourceEncoding=UTF-8
 sonar.verbose=true
 
-sonar.sources=src/
+sonar.sources=lib/
 sonar.language=js
 
-sonar.tests=tests/
+sonar.test.inclusions=**/*.spec.js
 sonar.coverage.exclusions=**/*
